requirements.txt
# Telegram bot
python-telegram-bot==20.5

# NLP (русский + английский)
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
ru-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.7.0/ru_core_news_lg-3.7.0-py3-none-any.whl

# Работа с документами (PDF/DOCX)
python-docx==0.8.11
pdfplumber==0.9.0
pdfminer.six==20221105

# Голосовой ввод/вывод
openai-whisper @ git+https://github.com/openai/whisper.git
torch==2.8.0
torchaudio==2.8.0
pyttsx3==2.90

# Yandex SpeechKit (альтернатива/дополнение к pyttsx3)
SpeechKit==1.3.1

# Генерация отчетов
fpdf==1.7.2
pandas>=1.0.0,<1.5.0
numpy==1.26.2

# Удобства
python-dotenv==1.0.1
requests==2.32.4
